Fix and re-enable readability-redundant-casting

+++ src/ra/nullmgr.cc
@@ -722,7 +722,7 @@ int NullModemClass::Service() {
   }
 
   RXCount += SerialPort->Read_From_Serial_Port(
-      (unsigned char*)(RXBuf + RXCount), static_cast<int>(RXSize - RXCount));
+      (unsigned char*)(RXBuf + RXCount), RXSize - RXCount);
 
   // minimum packet size
